+++ pialert/device.py
@@ -1,0 +1,434 @@
+
+
+
+
+import subprocess
+
+from pialert.conf import PHOLUS_ACTIVE, PHOLUS_FORCE, PHOLUS_TIMEOUT, cycle, DIG_GET_IP_ARG, userSubnets
+from pialert.helper import timeNow
+from pialert.internet import check_IP_format, get_internet_IP
+from pialert.logger import mylog, print_log
+from pialert.mac_vendor import query_MAC_vendor
+from pialert.pholusscan import performPholusScan, resolve_device_name_pholus
+#-------------------------------------------------------------------------------
+
+
+def save_scanned_devices (db, p_arpscan_devices, p_cycle_interval):
+    sql = db.sql #TO-DO
+    cycle = 1 # always 1, only one cycle supported
+
+    # Delete previous scan data
+    sql.execute ("DELETE FROM CurrentScan WHERE cur_ScanCycle = ?",
+                (cycle,))
+
+    if len(p_arpscan_devices) > 0:
+        # Insert new arp-scan devices
+        sql.executemany ("INSERT INTO CurrentScan (cur_ScanCycle, cur_MAC, "+
+                        "    cur_IP, cur_Vendor, cur_ScanMethod) "+
+                        "VALUES ("+ str(cycle) + ", :mac, :ip, :hw, 'arp-scan')",
+                        p_arpscan_devices) 
+
+    # Insert Pi-hole devices
+    startTime = timeNow()
+    sql.execute ("""INSERT INTO CurrentScan (cur_ScanCycle, cur_MAC, 
+                        cur_IP, cur_Vendor, cur_ScanMethod)
+                    SELECT ?, PH_MAC, PH_IP, PH_Vendor, 'Pi-hole'
+                    FROM PiHole_Network
+                    WHERE PH_LastQuery >= ?
+                      AND NOT EXISTS (SELECT 'X' FROM CurrentScan
+                                      WHERE cur_MAC = PH_MAC
+                                        AND cur_ScanCycle = ? )""",
+                    (cycle,
+                     (int(startTime.strftime('%s')) - 60 * p_cycle_interval),
+                     cycle) )
+
+    # Check Internet connectivity
+    internet_IP = get_internet_IP(DIG_GET_IP_ARG)
+        # TESTING - Force IP
+        # internet_IP = ""
+    if internet_IP != "" :
+        sql.execute ("""INSERT INTO CurrentScan (cur_ScanCycle, cur_MAC, cur_IP, cur_Vendor, cur_ScanMethod)
+                        VALUES (?, 'Internet', ?, Null, 'queryDNS') """, (cycle, internet_IP) )
+
+    # #76 Add Local MAC of default local interface
+      # BUGFIX #106 - Device that pialert is running
+        # local_mac_cmd = ["bash -lc ifconfig `ip route list default | awk {'print $5'}` | grep ether | awk '{print $2}'"]
+          # local_mac_cmd = ["/sbin/ifconfig `ip route list default | sort -nk11 | head -1 | awk {'print $5'}` | grep ether | awk '{print $2}'"]
+    local_mac_cmd = ["/sbin/ifconfig `ip -o route get 1 | sed 's/^.*dev \\([^ ]*\\).*$/\\1/;q'` | grep ether | awk '{print $2}'"]
+    local_mac = subprocess.Popen (local_mac_cmd, shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT).communicate()[0].decode().strip()
+
+    # local_dev_cmd = ["ip -o route get 1 | sed 's/^.*dev \\([^ ]*\\).*$/\\1/;q'"]
+    # local_dev = subprocess.Popen (local_dev_cmd, shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT).communicate()[0].decode().strip()
+    
+    # local_ip_cmd = ["ip route list default | awk {'print $7'}"]
+    local_ip_cmd = ["ip -o route get 1 | sed 's/^.*src \\([^ ]*\\).*$/\\1/;q'"]
+    local_ip = subprocess.Popen (local_ip_cmd, shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT).communicate()[0].decode().strip()
+
+    mylog('debug', ['    Saving this IP into the CurrentScan table:', local_ip])
+
+    if check_IP_format(local_ip) == '':
+        local_ip = '0.0.0.0'
+
+    # Check if local mac has been detected with other methods
+    sql.execute ("SELECT COUNT(*) FROM CurrentScan WHERE cur_ScanCycle = ? AND cur_MAC = ? ", (cycle, local_mac) )
+    if sql.fetchone()[0] == 0 :
+        sql.execute ("INSERT INTO CurrentScan (cur_ScanCycle, cur_MAC, cur_IP, cur_Vendor, cur_ScanMethod) "+
+                     "VALUES ( ?, ?, ?, Null, 'local_MAC') ", (cycle, local_mac, local_ip) )
+
+#-------------------------------------------------------------------------------
+def print_scan_stats (db):
+    sql = db.sql #TO-DO
+    # Devices Detected
+    sql.execute ("""SELECT COUNT(*) FROM CurrentScan
+                    WHERE cur_ScanCycle = ? """,
+                    (cycle,))
+    mylog('verbose', ['    Devices Detected.......: ', str (sql.fetchone()[0]) ])
+
+    # Devices arp-scan
+    sql.execute ("""SELECT COUNT(*) FROM CurrentScan
+                    WHERE cur_ScanMethod='arp-scan' AND cur_ScanCycle = ? """,
+                    (cycle,))
+    mylog('verbose', ['        arp-scan detected..: ', str (sql.fetchone()[0]) ])
+
+    # Devices Pi-hole
+    sql.execute ("""SELECT COUNT(*) FROM CurrentScan
+                    WHERE cur_ScanMethod='PiHole' AND cur_ScanCycle = ? """,
+                    (cycle,))
+    mylog('verbose', ['        Pi-hole detected...: +' + str (sql.fetchone()[0]) ])
+
+    # New Devices
+    sql.execute ("""SELECT COUNT(*) FROM CurrentScan
+                    WHERE cur_ScanCycle = ? 
+                      AND NOT EXISTS (SELECT 1 FROM Devices
+                                      WHERE dev_MAC = cur_MAC) """,
+                    (cycle,))
+    mylog('verbose', ['        New Devices........: ' + str (sql.fetchone()[0]) ])
+
+    # Devices in this ScanCycle
+    sql.execute ("""SELECT COUNT(*) FROM Devices, CurrentScan
+                    WHERE dev_MAC = cur_MAC AND dev_ScanCycle = cur_ScanCycle
+                      AND dev_ScanCycle = ? """,
+                    (cycle,))
+    
+    mylog('verbose', ['    Devices in this cycle..: ' + str (sql.fetchone()[0]) ])
+
+    # Down Alerts
+    sql.execute ("""SELECT COUNT(*) FROM Devices
+                    WHERE dev_AlertDeviceDown = 1
+                      AND dev_ScanCycle = ?
+                      AND NOT EXISTS (SELECT 1 FROM CurrentScan
+                                      WHERE dev_MAC = cur_MAC
+                                        AND dev_ScanCycle = cur_ScanCycle) """,
+                    (cycle,))
+    mylog('verbose', ['        Down Alerts........: ' + str (sql.fetchone()[0]) ])
+
+    # New Down Alerts
+    sql.execute ("""SELECT COUNT(*) FROM Devices
+                    WHERE dev_AlertDeviceDown = 1
+                      AND dev_PresentLastScan = 1
+                      AND dev_ScanCycle = ?
+                      AND NOT EXISTS (SELECT 1 FROM CurrentScan
+                                      WHERE dev_MAC = cur_MAC
+                                        AND dev_ScanCycle = cur_ScanCycle) """,
+                    (cycle,))
+    mylog('verbose', ['        New Down Alerts....: ' + str (sql.fetchone()[0]) ])
+
+    # New Connections
+    sql.execute ("""SELECT COUNT(*) FROM Devices, CurrentScan
+                    WHERE dev_MAC = cur_MAC AND dev_ScanCycle = cur_ScanCycle
+                      AND dev_PresentLastScan = 0
+                      AND dev_ScanCycle = ? """,
+                    (cycle,))
+    mylog('verbose', ['        New Connections....: ' + str ( sql.fetchone()[0]) ])
+
+    # Disconnections
+    sql.execute ("""SELECT COUNT(*) FROM Devices
+                    WHERE dev_PresentLastScan = 1
+                      AND dev_ScanCycle = ?
+                      AND NOT EXISTS (SELECT 1 FROM CurrentScan
+                                      WHERE dev_MAC = cur_MAC
+                                        AND dev_ScanCycle = cur_ScanCycle) """,
+                    (cycle,))
+    mylog('verbose', ['        Disconnections.....: ' + str ( sql.fetchone()[0]) ])
+
+    # IP Changes
+    sql.execute ("""SELECT COUNT(*) FROM Devices, CurrentScan
+                    WHERE dev_MAC = cur_MAC AND dev_ScanCycle = cur_ScanCycle
+                      AND dev_ScanCycle = ?
+                      AND dev_LastIP <> cur_IP """,
+                    (cycle,))
+    mylog('verbose', ['        IP Changes.........: ' + str ( sql.fetchone()[0]) ])
+
+
+
+#-------------------------------------------------------------------------------
+def create_new_devices (db):
+    sql = db.sql # TO-DO
+    startTime = timeNow()
+
+    # arpscan - Insert events for new devices
+    print_log ('New devices - 1 Events')
+    sql.execute ("""INSERT INTO Events (eve_MAC, eve_IP, eve_DateTime,
+                        eve_EventType, eve_AdditionalInfo,
+                        eve_PendingAlertEmail)
+                    SELECT cur_MAC, cur_IP, ?, 'New Device', cur_Vendor, 1
+                    FROM CurrentScan
+                    WHERE cur_ScanCycle = ? 
+                      AND NOT EXISTS (SELECT 1 FROM Devices
+                                      WHERE dev_MAC = cur_MAC) """,
+                    (startTime, cycle) ) 
+
+    print_log ('New devices - Insert Connection into session table')
+    sql.execute ("""INSERT INTO Sessions (ses_MAC, ses_IP, ses_EventTypeConnection, ses_DateTimeConnection,
+                        ses_EventTypeDisconnection, ses_DateTimeDisconnection, ses_StillConnected, ses_AdditionalInfo)
+                    SELECT cur_MAC, cur_IP,'Connected',?, NULL , NULL ,1, cur_Vendor
+                    FROM CurrentScan 
+                    WHERE cur_ScanCycle = ? 
+                      AND NOT EXISTS (SELECT 1 FROM Sessions
+                                      WHERE ses_MAC = cur_MAC) """,
+                    (startTime, cycle) ) 
+                    
+    # arpscan - Create new devices
+    print_log ('New devices - 2 Create devices')
+    sql.execute ("""INSERT INTO Devices (dev_MAC, dev_name, dev_Vendor,
+                        dev_LastIP, dev_FirstConnection, dev_LastConnection,
+                        dev_ScanCycle, dev_AlertEvents, dev_AlertDeviceDown,
+                        dev_PresentLastScan)
+                    SELECT cur_MAC, '(unknown)', cur_Vendor, cur_IP, ?, ?,
+                        1, 1, 0, 1
+                    FROM CurrentScan
+                    WHERE cur_ScanCycle = ? 
+                      AND NOT EXISTS (SELECT 1 FROM Devices
+                                      WHERE dev_MAC = cur_MAC) """,
+                    (startTime, startTime, cycle) ) 
+
+    # Pi-hole - Insert events for new devices
+    # NOT STRICYLY NECESARY (Devices can be created through Current_Scan)
+    # Bugfix #2 - Pi-hole devices w/o IP
+    print_log ('New devices - 3 Pi-hole Events')
+    sql.execute ("""INSERT INTO Events (eve_MAC, eve_IP, eve_DateTime,
+                        eve_EventType, eve_AdditionalInfo,
+                        eve_PendingAlertEmail)
+                    SELECT PH_MAC, IFNULL (PH_IP,'-'), ?, 'New Device',
+                        '(Pi-Hole) ' || PH_Vendor, 1
+                    FROM PiHole_Network
+                    WHERE NOT EXISTS (SELECT 1 FROM Devices
+                                      WHERE dev_MAC = PH_MAC) """,
+                    (startTime, ) ) 
+
+    # Pi-hole - Create New Devices
+    # Bugfix #2 - Pi-hole devices w/o IP
+    print_log ('New devices - 4 Pi-hole Create devices')
+    sql.execute ("""INSERT INTO Devices (dev_MAC, dev_name, dev_Vendor,
+                        dev_LastIP, dev_FirstConnection, dev_LastConnection,
+                        dev_ScanCycle, dev_AlertEvents, dev_AlertDeviceDown,
+                        dev_PresentLastScan)
+                    SELECT PH_MAC, PH_Name, PH_Vendor, IFNULL (PH_IP,'-'),
+                        ?, ?, 1, 1, 0, 1
+                    FROM PiHole_Network
+                    WHERE NOT EXISTS (SELECT 1 FROM Devices
+                                      WHERE dev_MAC = PH_MAC) """,
+                    (startTime, startTime) ) 
+
+    # DHCP Leases - Insert events for new devices
+    print_log ('New devices - 5 DHCP Leases Events')
+    sql.execute ("""INSERT INTO Events (eve_MAC, eve_IP, eve_DateTime,
+                        eve_EventType, eve_AdditionalInfo,
+                        eve_PendingAlertEmail)
+                    SELECT DHCP_MAC, DHCP_IP, ?, 'New Device', '(DHCP lease)',1
+                    FROM DHCP_Leases
+                    WHERE NOT EXISTS (SELECT 1 FROM Devices
+                                      WHERE dev_MAC = DHCP_MAC) """,
+                    (startTime, ) ) 
+
+    # DHCP Leases - Create New Devices
+    print_log ('New devices - 6 DHCP Leases Create devices')
+    # BUGFIX #23 - Duplicated MAC in DHCP.Leases
+    # TEST - Force Duplicated MAC
+        # sql.execute ("""INSERT INTO DHCP_Leases VALUES
+        #                 (1610700000, 'TEST1', '10.10.10.1', 'Test 1', '*')""")
+        # sql.execute ("""INSERT INTO DHCP_Leases VALUES
+        #                 (1610700000, 'TEST2', '10.10.10.2', 'Test 2', '*')""")
+    sql.execute ("""INSERT INTO Devices (dev_MAC, dev_name, dev_LastIP, 
+                        dev_Vendor, dev_FirstConnection, dev_LastConnection,
+                        dev_ScanCycle, dev_AlertEvents, dev_AlertDeviceDown,
+                        dev_PresentLastScan)
+                    SELECT DISTINCT DHCP_MAC,
+                        (SELECT DHCP_Name FROM DHCP_Leases AS D2
+                         WHERE D2.DHCP_MAC = D1.DHCP_MAC
+                         ORDER BY DHCP_DateTime DESC LIMIT 1),
+                        (SELECT DHCP_IP FROM DHCP_Leases AS D2
+                         WHERE D2.DHCP_MAC = D1.DHCP_MAC
+                         ORDER BY DHCP_DateTime DESC LIMIT 1),
+                        '(unknown)', ?, ?, 1, 1, 0, 1
+                    FROM DHCP_Leases AS D1
+                    WHERE NOT EXISTS (SELECT 1 FROM Devices
+                                      WHERE dev_MAC = DHCP_MAC) """,
+                    (startTime, startTime) ) 
+
+    # sql.execute ("""INSERT INTO Devices (dev_MAC, dev_name, dev_Vendor,
+    #                     dev_LastIP, dev_FirstConnection, dev_LastConnection,
+    #                     dev_ScanCycle, dev_AlertEvents, dev_AlertDeviceDown,
+    #                     dev_PresentLastScan)
+    #                 SELECT DHCP_MAC, DHCP_Name, '(unknown)', DHCP_IP, ?, ?,
+    #                     1, 1, 0, 1
+    #                 FROM DHCP_Leases
+    #                 WHERE NOT EXISTS (SELECT 1 FROM Devices
+    #                                   WHERE dev_MAC = DHCP_MAC) """,
+    #                 (startTime, startTime) ) 
+    print_log ('New Devices end')
+    db.commit()
+
+
+#-------------------------------------------------------------------------------
+def update_devices_data_from_scan (db):
+    sql = db.sql #TO-DO    
+    startTime = timeNow()
+    # Update Last Connection
+    print_log ('Update devices - 1 Last Connection')
+    sql.execute ("""UPDATE Devices SET dev_LastConnection = ?,
+                        dev_PresentLastScan = 1
+                    WHERE dev_ScanCycle = ?
+                      AND dev_PresentLastScan = 0
+                      AND EXISTS (SELECT 1 FROM CurrentScan 
+                                  WHERE dev_MAC = cur_MAC
+                                    AND dev_ScanCycle = cur_ScanCycle) """,
+                    (startTime, cycle))
+
+    # Clean no active devices
+    print_log ('Update devices - 2 Clean no active devices')
+    sql.execute ("""UPDATE Devices SET dev_PresentLastScan = 0
+                    WHERE dev_ScanCycle = ?
+                      AND NOT EXISTS (SELECT 1 FROM CurrentScan 
+                                      WHERE dev_MAC = cur_MAC
+                                        AND dev_ScanCycle = cur_ScanCycle) """,
+                    (cycle,))
+
+    # Update IP & Vendor
+    print_log ('Update devices - 3 LastIP & Vendor')
+    sql.execute ("""UPDATE Devices
+                    SET dev_LastIP = (SELECT cur_IP FROM CurrentScan
+                                      WHERE dev_MAC = cur_MAC
+                                        AND dev_ScanCycle = cur_ScanCycle),
+                        dev_Vendor = (SELECT cur_Vendor FROM CurrentScan
+                                      WHERE dev_MAC = cur_MAC
+                                        AND dev_ScanCycle = cur_ScanCycle)
+                    WHERE dev_ScanCycle = ?
+                      AND EXISTS (SELECT 1 FROM CurrentScan
+                                  WHERE dev_MAC = cur_MAC
+                                    AND dev_ScanCycle = cur_ScanCycle) """,
+                    (cycle,)) 
+
+    # Pi-hole Network - Update (unknown) Name
+    print_log ('Update devices - 4 Unknown Name')
+    sql.execute ("""UPDATE Devices
+                    SET dev_NAME = (SELECT PH_Name FROM PiHole_Network
+                                    WHERE PH_MAC = dev_MAC)
+                    WHERE (dev_Name in ("(unknown)", "(name not found)", "" )
+                           OR dev_Name IS NULL)
+                      AND EXISTS (SELECT 1 FROM PiHole_Network
+                                  WHERE PH_MAC = dev_MAC
+                                    AND PH_NAME IS NOT NULL
+                                    AND PH_NAME <> '') """)
+
+    # DHCP Leases - Update (unknown) Name
+    sql.execute ("""UPDATE Devices
+                    SET dev_NAME = (SELECT DHCP_Name FROM DHCP_Leases
+                                    WHERE DHCP_MAC = dev_MAC)
+                    WHERE (dev_Name in ("(unknown)", "(name not found)", "" )                           
+                           OR dev_Name IS NULL)
+                      AND EXISTS (SELECT 1 FROM DHCP_Leases
+                                  WHERE DHCP_MAC = dev_MAC)""")
+
+    # DHCP Leases - Vendor
+    print_log ('Update devices - 5 Vendor')
+
+    recordsToUpdate = []
+    query = """SELECT * FROM Devices
+               WHERE dev_Vendor = '(unknown)' OR dev_Vendor =''
+                  OR dev_Vendor IS NULL"""
+
+    for device in sql.execute (query) :
+        vendor = query_MAC_vendor (device['dev_MAC'])
+        if vendor != -1 and vendor != -2 :
+            recordsToUpdate.append ([vendor, device['dev_MAC']])
+
+    sql.executemany ("UPDATE Devices SET dev_Vendor = ? WHERE dev_MAC = ? ",
+        recordsToUpdate )
+
+    # clean-up device leases table
+    sql.execute ("DELETE FROM DHCP_Leases")
+    print_log ('Update devices end')
+
+#-------------------------------------------------------------------------------
+def update_devices_names (db):
+    sql = db.sql #TO-DO
+    # Initialize variables
+    recordsToUpdate = []
+    recordsNotFound = []
+
+    ignored = 0
+    notFound = 0
+
+    foundDig = 0
+    foundPholus = 0
+
+    # BUGFIX #97 - Updating name of Devices w/o IP    
+    sql.execute ("SELECT * FROM Devices WHERE dev_Name IN ('(unknown)','', '(name not found)') AND dev_LastIP <> '-'")
+    unknownDevices = sql.fetchall() 
+    db.commitDB()
+
+    # perform Pholus scan if (unknown) devices found
+    if PHOLUS_ACTIVE and (len(unknownDevices) > 0 or PHOLUS_FORCE):        
+        performPholusScan(db, PHOLUS_TIMEOUT, userSubnets)
+
+    # skip checks if no unknown devices
+    if len(unknownDevices) == 0 and PHOLUS_FORCE == False:
+        return
+
+    # Devices without name
+    mylog('verbose', ['        Trying to resolve devices without name'])
+
+    # get names from Pholus scan 
+    sql.execute ('SELECT * FROM Pholus_Scan where "Record_Type"="Answer"')    
+    pholusResults = list(sql.fetchall())        
+    db.commitDB()
+
+    # Number of entries from previous Pholus scans
+    mylog('verbose', ["          Pholus entries from prev scans: ", len(pholusResults)])
+
+    for device in unknownDevices:
+        newName = -1
+        
+        # Resolve device name with DiG
+        newName = resolve_device_name_pholus (device['dev_MAC'], device['dev_LastIP'])
+        
+        # count
+        if newName != -1:
+            foundDig += 1
+
+        # Resolve with Pholus 
+        if newName == -1:
+            newName =  resolve_device_name_pholus (device['dev_MAC'], device['dev_LastIP'], pholusResults)
+            # count
+            if newName != -1:
+                foundPholus += 1
+        
+        # isf still not found update name so we can distinguish the devices where we tried already
+        if newName == -1 :
+            recordsNotFound.append (["(name not found)", device['dev_MAC']])          
+        else:
+            # name wa sfound with DiG or Pholus
+            recordsToUpdate.append ([newName, device['dev_MAC']])
+
+    # Print log            
+    mylog('verbose', ["        Names Found (DiG/Pholus): ", len(recordsToUpdate), " (",foundDig,"/",foundPholus ,")" ])                 
+    mylog('verbose', ["        Names Not Found         : ", len(recordsNotFound) ])    
+     
+    # update not found devices with (name not found) 
+    sql.executemany ("UPDATE Devices SET dev_Name = ? WHERE dev_MAC = ? ", recordsNotFound )
+    # update names of devices which we were bale to resolve
+    sql.executemany ("UPDATE Devices SET dev_Name = ? WHERE dev_MAC = ? ", recordsToUpdate )
+    db.commitDB()
+
+
